Restructure into individual classes
